Refactor `GraduateProgramRequirementsSummary` to `GraduateProgramSummary`, update layout to use `getGraduateProgramById` data, and simplify type definitions.
